remove criterion harness for raft log store bench
